cap: move handleCAP function from CommandHandler.cpp to handleCAP.cpp. Adjust Makefile
